flask-app/app/config.py
"""
Configuration de l'application
"""
import os
from dotenv import load_dotenv

# Charger les variables d'environnement
load_dotenv()

class Config:
    """Configuration de l'application Flask"""
    
    # Configuration Flask
    SECRET_KEY = os.environ.get('JWT_SECRET_KEY')
    
    # Configuration Base de données
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    
    # Configuration JWT
    JWT_SECRET_KEY = os.environ.get('JWT_SECRET_KEY')
    JWT_ACCESS_TOKEN_EXPIRES = 24 * 3600  # 24 heures en secondes 