flask-app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
python-dotenv==0.19.0
PyMySQL==1.0.2
bcrypt==3.2.0
PyJWT==2.3.0
cryptography==3.4.7
Werkzeug==2.2.3
requests==2.28.1
SQLAlchemy>=1.4.0,<2.0.0